[IMP] base: possibility to add pdf-engine such as Paper-muncher
