+++ addons/web/static/tests/webclient/actions/report_action.test.js
@@ -97,7 +97,7 @@ test("can execute report actions from db ID", async () => {
             return Promise.resolve();
         },
     });
-    onRpc("/report/check_wkhtmltopdf", () => "ok");
+    onRpc("/report/check_pdf_engine", () => ["wkhtmltopdf", "ok"]);
     stepAllNetworkCalls();
 
     await mountWithCleanup(WebClient);
@@ -106,7 +106,7 @@ test("can execute report actions from db ID", async () => {
         "/web/webclient/translations",
         "/web/webclient/load_menus",
         "/web/action/load",
-        "/report/check_wkhtmltopdf",
+        "/report/check_pdf_engine",
         "/report/download",
         "on_close",
     ]);
@@ -129,7 +129,7 @@ test("report actions can close modals and reload views", async () => {
         },
     });
 
-    onRpc("/report/check_wkhtmltopdf", () => "ok");
+    onRpc("/report/check_pdf_engine", () => ["wkhtmltopdf", "ok"]);
 
     await mountWithCleanup(WebClient);
     await getService("action").doAction(5, { onClose: () => expect.step("on_close") });
@@ -160,7 +160,7 @@ test("should trigger a notification if wkhtmltopdf is to upgrade", async () => {
         add: () => expect.step("notify"),
     });
 
-    onRpc("/report/check_wkhtmltopdf", () => "upgrade");
+    onRpc("/report/check_pdf_engine", () => ["wkhtmltopdf", "upgrade"]);
     stepAllNetworkCalls();
 
     await mountWithCleanup(WebClient);
@@ -169,7 +169,7 @@ test("should trigger a notification if wkhtmltopdf is to upgrade", async () => {
         "/web/webclient/translations",
         "/web/webclient/load_menus",
         "/web/action/load",
-        "/report/check_wkhtmltopdf",
+        "/report/check_pdf_engine",
         "/report/download",
         "notify",
     ]);
@@ -195,7 +195,7 @@ test("should open the report client action if wkhtmltopdf is broken", async () =
         add: () => expect.step("notify"),
     });
 
-    onRpc("/report/check_wkhtmltopdf", () => "broken");
+    onRpc("/report/check_pdf_engine", () => ["wkhtmltopdf", "broken"]);
     onRpc("/report/html/some_report", async (request) => {
         const search = decodeURIComponent(new URL(request.url).search);
         expect(search).toBe(`?context={"lang":"en","tz":"taht","uid":7,"allowed_company_ids":[1]}`);
@@ -216,7 +216,7 @@ test("should open the report client action if wkhtmltopdf is broken", async () =
         "/web/webclient/translations",
         "/web/webclient/load_menus",
         "/web/action/load",
-        "/report/check_wkhtmltopdf",
+        "/report/check_pdf_engine",
         "notify",
         "/report/html/some_report",
     ]);
@@ -281,7 +281,7 @@ test("UI unblocks after downloading the report even if it threw an error", async
         },
     });
 
-    onRpc("/report/check_wkhtmltopdf", () => "ok");
+    onRpc("/report/check_pdf_engine", () => ["wkhtmltopdf", "ok"]);
 
     await mountWithCleanup(WebClient);
     const onBlock = () => {
@@ -320,7 +320,7 @@ test("can use custom handlers for report actions", async () => {
         },
     });
 
-    onRpc("/report/check_wkhtmltopdf", () => "ok");
+    onRpc("/report/check_pdf_engine", () => ["wkhtmltopdf", "ok"]);
     stepAllNetworkCalls();
 
     await mountWithCleanup(WebClient);
@@ -344,7 +344,7 @@ test("can use custom handlers for report actions", async () => {
         "calling custom handler",
         "first doAction finished",
         "falling through to default handler",
-        "/report/check_wkhtmltopdf",
+        "/report/check_pdf_engine",
         "/report/download",
     ]);
 });
@@ -372,7 +372,7 @@ test("context is correctly passed to the client action report", async (assert) =
         },
     });
 
-    onRpc("/report/check_wkhtmltopdf", () => "ok");
+    onRpc("/report/check_pdf_engine", () => ["wkhtmltopdf", "ok"]);
     onRpc("/report/html", async (request) => {
         const search = decodeURIComponent(new URL(request.url).search);
         expect(search).toBe(`?context={"lang":"en","tz":"taht","uid":7,"allowed_company_ids":[1]}`);
@@ -399,7 +399,7 @@ test("context is correctly passed to the client action report", async (assert) =
     expect.verifySteps(["/report/html/ennio.morricone/99"]);
 
     await contains(".o_control_panel_main_buttons button[title='Print']").click();
-    expect.verifySteps(["/report/check_wkhtmltopdf", "/report/download"]);
+    expect.verifySteps(["/report/check_pdf_engine", "/report/download"]);
 });
 
 test("url is valid", async (assert) => {
