+++ odoo/addons/base/models/__init__.py
@@ -10,6 +10,8 @@ from . import ir_asset
 from . import ir_actions
 from . import ir_embedded_actions
 from . import ir_actions_report
+from . import ir_actions_report_wkhtmltopdf
+from . import ir_actions_report_papermuncher
 from . import ir_attachment
 from . import ir_binary
 from . import ir_cron
