+++ odoo/addons/base/models/ir_actions_report.py
@@ -3,17 +3,9 @@
 import io
 import json
 import logging
-import os
-import re
-import subprocess
-import tempfile
-import unittest
 from ast import literal_eval
 from collections import OrderedDict
 from collections.abc import Iterable
-from contextlib import closing, ExitStack
-from itertools import islice
-from urllib.parse import urlparse
 
 import lxml.html
 from PIL import Image, ImageFile
@@ -24,11 +16,8 @@ from reportlab.pdfbase.pdfmetrics import getFont, TypeFace
 
 from odoo import api, fields, models, modules, tools, _
 from odoo.exceptions import UserError, AccessError, RedirectWarning
-from odoo.service import security
-from odoo.http import request, root
 from odoo.osv.expression import NEGATIVE_TERM_OPERATORS, FALSE_DOMAIN
-from odoo.tools import check_barcode_encoding, config, is_html_empty, parse_version, split_every
-from odoo.tools.misc import find_in_path
+from odoo.tools import check_barcode_encoding, is_html_empty
 from odoo.tools.pdf import PdfFileReader, PdfFileWriter, PdfReadError
 from odoo.tools.safe_eval import safe_eval, time
 
@@ -56,84 +45,6 @@ try:
 except Exception:
     pass
 
-
-def _get_wkhtmltopdf_bin():
-    return find_in_path('wkhtmltopdf')
-
-
-def _get_wkhtmltoimage_bin():
-    return find_in_path('wkhtmltoimage')
-
-
-def _split_table(tree, max_rows):
-    """
-    Walks through the etree and splits tables with more than max_rows rows into
-    multiple tables with max_rows rows.
-
-    This function is needed because wkhtmltopdf has a exponential processing
-    time growth when processing tables with many rows. This function is a
-    workaround for this problem.
-
-    :param tree: The etree to process
-    :param max_rows: The maximum number of rows per table
-    """
-    for table in list(tree.iter('table')):
-        prev = table
-        for rows in islice(split_every(max_rows, table), 1, None):
-            sibling = etree.Element('table', attrib=table.attrib)
-            sibling.extend(rows)
-            prev.addnext(sibling)
-            prev = sibling
-
-# Check the presence of Wkhtmltopdf and return its version at Odoo start-up
-wkhtmltopdf_state = 'install'
-wkhtmltopdf_dpi_zoom_ratio = False
-try:
-    process = subprocess.Popen(
-        [_get_wkhtmltopdf_bin(), '--version'], stdout=subprocess.PIPE, stderr=subprocess.PIPE
-    )
-except (OSError, IOError):
-    _logger.info('You need Wkhtmltopdf to print a pdf version of the reports.')
-else:
-    _logger.info('Will use the Wkhtmltopdf binary at %s' % _get_wkhtmltopdf_bin())
-    out, err = process.communicate()
-    match = re.search(b'([0-9.]+)', out)
-    if match:
-        version = match.group(0).decode('ascii')
-        if parse_version(version) < parse_version('0.12.0'):
-            _logger.info('Upgrade Wkhtmltopdf to (at least) 0.12.0')
-            wkhtmltopdf_state = 'upgrade'
-        else:
-            wkhtmltopdf_state = 'ok'
-        if parse_version(version) >= parse_version('0.12.2'):
-            wkhtmltopdf_dpi_zoom_ratio = True
-
-        if config['workers'] == 1:
-            _logger.info('You need to start Odoo with at least two workers to print a pdf version of the reports.')
-            wkhtmltopdf_state = 'workers'
-    else:
-        _logger.info('Wkhtmltopdf seems to be broken.')
-        wkhtmltopdf_state = 'broken'
-
-wkhtmltoimage_version = None
-try:
-    process = subprocess.Popen(
-        [_get_wkhtmltoimage_bin(), '--version'], stdout=subprocess.PIPE, stderr=subprocess.PIPE
-    )
-except OSError:
-    _logger.info('You need Wkhtmltoimage to generate images from html.')
-else:
-    _logger.info('Will use the Wkhtmltoimage binary at %s', _get_wkhtmltoimage_bin())
-    out, err = process.communicate()
-    match = re.search(b'([0-9.]+)', out)
-    if match:
-        wkhtmltoimage_version = parse_version(match.group(0).decode('ascii'))
-        if config['workers'] == 1:
-            _logger.info('You need to start Odoo with at least two workers to convert images to html.')
-    else:
-        _logger.info('Wkhtmltoimage seems to be broken.')
-
-
 class IrActionsReport(models.Model):
     _name = 'ir.actions.report'
     _description = 'Report Action'
@@ -151,10 +62,10 @@ class IrActionsReport(models.Model):
         ('qweb-html', 'HTML'),
         ('qweb-pdf', 'PDF'),
         ('qweb-text', 'Text'),
-    ], required=True, default='qweb-pdf',
+    ], required=True, default='qweb-html',
     help='The type of the report that will be rendered, each one having its own'
         ' rendering method. HTML means the report will be opened directly in your'
-        ' browser PDF means the report will be rendered using Wkhtmltopdf and'
+        ' browser PDF means the report will be rendered using a pdf engine and'
         ' downloaded by the user.')
     report_name = fields.Char(string='Template Name', required=True)
     report_file = fields.Char(string='Report File', required=False, readonly=False, store=True,
@@ -256,19 +167,6 @@ class IrActionsReport(models.Model):
                 ('res_id', '=', record.id)
         ], limit=1)
 
-    @api.model
-    def get_wkhtmltopdf_state(self):
-        '''Get the current state of wkhtmltopdf: install, ok, upgrade, workers or broken.
-        * install: Starting state.
-        * upgrade: The binary is an older version (< 0.12.0).
-        * ok: A binary was found with a recent version (>= 0.12.0).
-        * workers: Not enough workers found to perform the pdf rendering process (< 2 workers).
-        * broken: A binary was found but not responding.
-
-        :return: wkhtmltopdf_state
-        '''
-        return wkhtmltopdf_state
-
     def get_paperformat(self):
         return self.paperformat_id or self.env.company.paperformat_id
 
@@ -281,88 +179,6 @@ class IrActionsReport(models.Model):
     def _get_report_url(self, layout=None):
         report_url = self.env['ir.config_parameter'].sudo().get_param('report.url')
         return report_url or (layout or self._get_layout() or self).get_base_url()
-
-    @api.model
-    def _build_wkhtmltopdf_args(
-            self,
-            paperformat_id,
-            landscape,
-            specific_paperformat_args=None,
-            set_viewport_size=False):
-        '''Build arguments understandable by wkhtmltopdf bin.
-
-        :param paperformat_id: A report.paperformat record.
-        :param landscape: Force the report orientation to be landscape.
-        :param specific_paperformat_args: A dictionary containing prioritized wkhtmltopdf arguments.
-        :param set_viewport_size: Enable a viewport sized '1024x1280' or '1280x1024' depending of landscape arg.
-        :return: A list of string representing the wkhtmltopdf process command args.
-        '''
-        if landscape is None and specific_paperformat_args and specific_paperformat_args.get('data-report-landscape'):
-            landscape = specific_paperformat_args.get('data-report-landscape')
-
-        command_args = ['--disable-local-file-access']
-        if set_viewport_size:
-            command_args.extend(['--viewport-size', landscape and '1024x1280' or '1280x1024'])
-
-        # Less verbose error messages
-        command_args.extend(['--quiet'])
-
-        # Build paperformat args
-        if paperformat_id:
-            if paperformat_id.format and paperformat_id.format != 'custom':
-                command_args.extend(['--page-size', paperformat_id.format])
-
-            if paperformat_id.page_height and paperformat_id.page_width and paperformat_id.format == 'custom':
-                command_args.extend(['--page-width', str(paperformat_id.page_width) + 'mm'])
-                command_args.extend(['--page-height', str(paperformat_id.page_height) + 'mm'])
-
-            if specific_paperformat_args and specific_paperformat_args.get('data-report-margin-top'):
-                command_args.extend(['--margin-top', str(specific_paperformat_args['data-report-margin-top'])])
-            else:
-                command_args.extend(['--margin-top', str(paperformat_id.margin_top)])
-
-            dpi = None
-            if specific_paperformat_args and specific_paperformat_args.get('data-report-dpi'):
-                dpi = int(specific_paperformat_args['data-report-dpi'])
-            elif paperformat_id.dpi:
-                if os.name == 'nt' and int(paperformat_id.dpi) <= 95:
-                    _logger.info("Generating PDF on Windows platform require DPI >= 96. Using 96 instead.")
-                    dpi = 96
-                else:
-                    dpi = paperformat_id.dpi
-            if dpi:
-                command_args.extend(['--dpi', str(dpi)])
-                if wkhtmltopdf_dpi_zoom_ratio:
-                    command_args.extend(['--zoom', str(96.0 / dpi)])
-
-            if specific_paperformat_args and specific_paperformat_args.get('data-report-header-spacing'):
-                command_args.extend(['--header-spacing', str(specific_paperformat_args['data-report-header-spacing'])])
-            elif paperformat_id.header_spacing:
-                command_args.extend(['--header-spacing', str(paperformat_id.header_spacing)])
-
-            command_args.extend(['--margin-left', str(paperformat_id.margin_left)])
-
-            if specific_paperformat_args and specific_paperformat_args.get('data-report-margin-bottom'):
-                command_args.extend(['--margin-bottom', str(specific_paperformat_args['data-report-margin-bottom'])])
-            else:
-                command_args.extend(['--margin-bottom', str(paperformat_id.margin_bottom)])
-
-            command_args.extend(['--margin-right', str(paperformat_id.margin_right)])
-            if not landscape and paperformat_id.orientation:
-                command_args.extend(['--orientation', str(paperformat_id.orientation)])
-            if paperformat_id.header_line:
-                command_args.extend(['--header-line'])
-            if paperformat_id.disable_shrinking:
-                command_args.extend(['--disable-smart-shrinking'])
-
-        # Add extra time to allow the page to render
-        delay = self.env['ir.config_parameter'].sudo().get_param('report.print_delay', '1000')
-        command_args.extend(['--javascript-delay', delay])
-
-        if landscape:
-            command_args.extend(['--orientation', 'landscape'])
-
-        return command_args
 
     def _prepare_html(self, html, report_model=False):
         '''Divide and recreate the header/footer html by merging all found in html.
@@ -449,186 +265,6 @@ class IrActionsReport(models.Model):
         })
 
         return bodies, res_ids, header, footer, specific_paperformat_args
-
-    def _run_wkhtmltoimage(self, bodies, width, height, image_format="jpg"):
-        """
-        :bodies str: valid html documents as strings
-        :param width int: width in pixels
-        :param height int: height in pixels
-        :param image_format union['jpg', 'png']: format of the image
-        :return list[bytes|None]:
-        """
-        if (modules.module.current_test or tools.config['test_enable']) and not self.env.context.get('force_image_rendering'):
-            return [None] * len(bodies)
-        if not wkhtmltoimage_version or wkhtmltoimage_version < parse_version('0.12.0'):
-            raise UserError(_('wkhtmltoimage 0.12.0^ is required in order to render images from html'))
-        command_args = [
-            '--disable-local-file-access', '--disable-javascript',
-            '--quiet',
-            '--width', str(width), '--height', str(height),
-            '--format', image_format,
-        ]
-        with ExitStack() as stack:
-            files = []
-            for body in bodies:
-                input_file = stack.enter_context(tempfile.NamedTemporaryFile(suffix='.html', prefix='report_image_html_input.tmp.'))
-                output_file = stack.enter_context(tempfile.NamedTemporaryFile(suffix=f'.{image_format}', prefix='report_image_output.tmp.'))
-                input_file.write(body.encode())
-                files.append((input_file, output_file))
-            output_images = []
-            for input_file, output_file in files:
-                # smaller bodies may be held in a python buffer until close, force flush
-                input_file.flush()
-                wkhtmltoimage = [_get_wkhtmltoimage_bin()] + command_args + [input_file.name, output_file.name]
-                # start and block, no need for parallelism for now
-                completed_process = subprocess.run(wkhtmltoimage, stdout=subprocess.DEVNULL, stderr=subprocess.PIPE, check=False)
-                if completed_process.returncode:
-                    message = _(
-                        'Wkhtmltoimage failed (error code: %(error_code)s). Message: %(error_message_end)s',
-                        error_code=completed_process.returncode,
-                        error_message_end=completed_process.stderr[-1000:],
-                    )
-                    _logger.warning(message)
-                    output_images.append(None)
-                else:
-                    output_images.append(output_file.read())
-        return output_images
-
-    @api.model
-    def _run_wkhtmltopdf(
-            self,
-            bodies,
-            report_ref=False,
-            header=None,
-            footer=None,
-            landscape=False,
-            specific_paperformat_args=None,
-            set_viewport_size=False):
-        '''Execute wkhtmltopdf as a subprocess in order to convert html given in input into a pdf
-        document.
-
-        :param list[str] bodies: The html bodies of the report, one per page.
-        :param report_ref: report reference that is needed to get report paperformat.
-        :param str header: The html header of the report containing all headers.
-        :param str footer: The html footer of the report containing all footers.
-        :param landscape: Force the pdf to be rendered under a landscape format.
-        :param specific_paperformat_args: dict of prioritized paperformat arguments.
-        :param set_viewport_size: Enable a viewport sized '1024x1280' or '1280x1024' depending of landscape arg.
-        :return: Content of the pdf as bytes
-        :rtype: bytes
-        '''
-        paperformat_id = self._get_report(report_ref).get_paperformat() if report_ref else self.get_paperformat()
-
-        # Build the base command args for wkhtmltopdf bin
-        command_args = self._build_wkhtmltopdf_args(
-            paperformat_id,
-            landscape,
-            specific_paperformat_args=specific_paperformat_args,
-            set_viewport_size=set_viewport_size)
-
-        files_command_args = []
-
-        def delete_file(file_path):
-            try:
-                os.unlink(file_path)
-            except OSError:
-                _logger.error('Error when trying to remove file %s', file_path)
-
-        with ExitStack() as stack:
-
-            # Passing the cookie to wkhtmltopdf in order to resolve internal links.
-            if request and request.db:
-                # Create a temporary session which will not create device logs
-                temp_session = root.session_store.new()
-                temp_session.update({
-                    **request.session,
-                    '_trace_disable': True,
-                })
-                if temp_session.uid:
-                    temp_session.session_token = security.compute_session_token(temp_session, self.env)
-                root.session_store.save(temp_session)
-                stack.callback(root.session_store.delete, temp_session)
-
-                base_url = self._get_report_url()
-                domain = urlparse(base_url).hostname
-                cookie = f'session_id={temp_session.sid}; HttpOnly; domain={domain}; path=/;'
-                cookie_jar_file_fd, cookie_jar_file_path = tempfile.mkstemp(suffix='.txt', prefix='report.cookie_jar.tmp.')
-                stack.callback(delete_file, cookie_jar_file_path)
-                with closing(os.fdopen(cookie_jar_file_fd, 'wb')) as cookie_jar_file:
-                    cookie_jar_file.write(cookie.encode())
-                command_args.extend(['--cookie-jar', cookie_jar_file_path])
-
-            if header:
-                head_file_fd, head_file_path = tempfile.mkstemp(suffix='.html', prefix='report.header.tmp.')
-                with closing(os.fdopen(head_file_fd, 'wb')) as head_file:
-                    head_file.write(header.encode())
-                stack.callback(delete_file, head_file_path)
-                files_command_args.extend(['--header-html', head_file_path])
-            if footer:
-                foot_file_fd, foot_file_path = tempfile.mkstemp(suffix='.html', prefix='report.footer.tmp.')
-                with closing(os.fdopen(foot_file_fd, 'wb')) as foot_file:
-                    foot_file.write(footer.encode())
-                stack.callback(delete_file, foot_file_path)
-                files_command_args.extend(['--footer-html', foot_file_path])
-
-            paths = []
-            for i, body in enumerate(bodies):
-                prefix = '%s%d.' % ('report.body.tmp.', i)
-                body_file_fd, body_file_path = tempfile.mkstemp(suffix='.html', prefix=prefix)
-                with closing(os.fdopen(body_file_fd, 'wb')) as body_file:
-                    # HACK: wkhtmltopdf doesn't like big table at all and the
-                    #       processing time become exponential with the number
-                    #       of rows (like 1H for 250k rows).
-                    #
-                    #       So we split the table into multiple tables containing
-                    #       500 rows each. This reduce the processing time to 1min
-                    #       for 250k rows. The number 500 was taken from opw-1689673
-                    if len(body) < 4 * 1024 * 1024:  # 4Mib
-                        body_file.write(body.encode())
-                    else:
-                        tree = lxml.html.fromstring(body)
-                        _split_table(tree, 500)
-                        body_file.write(lxml.html.tostring(tree))
-                paths.append(body_file_path)
-                stack.callback(delete_file, body_file_path)
-
-            pdf_report_fd, pdf_report_path = tempfile.mkstemp(suffix='.pdf', prefix='report.tmp.')
-            os.close(pdf_report_fd)
-            stack.callback(delete_file, pdf_report_path)
-
-            wkhtmltopdf = [_get_wkhtmltopdf_bin()] + command_args + files_command_args + paths + [pdf_report_path]
-            process = subprocess.Popen(wkhtmltopdf, stdout=subprocess.PIPE, stderr=subprocess.PIPE, encoding="utf-8")
-            _out, err = process.communicate()
-
-            match process.returncode:
-                case 0:
-                    pass
-                case 1:
-                    if len(bodies) > 1:
-                        v = subprocess.run([_get_wkhtmltopdf_bin(), '--version'], stdout=subprocess.PIPE, check=True, encoding="utf-8")
-                        if '(with patched qt)' not in v.stdout:
-                            if modules.module.current_test or tools.config['test_enable']:
-                                raise unittest.SkipTest("Unable to convert multiple documents via wkhtmltopdf using unpatched QT")
-                            raise UserError(_("Tried to convert multiple documents in wkhtmltopdf using unpatched QT"))
-
-                    _logger.warning("wkhtmltopdf: %s", err)
-                case c:
-                    message = _(
-                        'Wkhtmltopdf failed (error code: %(error_code)s). Memory limit too low or maximum file number of subprocess reached. Message : %(message)s',
-                        error_code=c,
-                        message=err[-1000:],
-                    ) if c == -11 else _(
-                        'Wkhtmltopdf failed (error code: %(error_code)s). Message: %(message)s',
-                        error_code=c,
-                        message=err[-1000:],
-                    )
-                    _logger.warning(message)
-                    raise UserError(message)
-
-            with open(pdf_report_path, 'rb') as pdf_document:
-                pdf_content = pdf_document.read()
-
-        return pdf_content
 
     @api.model
     def _get_report_from_name(self, report_name):
@@ -781,6 +417,39 @@ class IrActionsReport(models.Model):
         writer.write(result_stream)
         return result_stream
 
+    @api.model
+    def get_pdf_engine_state(self, engine_name=None):
+        """
+        Returns the default functional engine, or the requested engine status
+        """
+        raise NotImplementedError()
+
+    @api.model
+    def _get_render_pdf_engine(self, engine_name=None):
+        """
+        Return the callable method "_run_pdf_engine_*"
+        depending on the installed libraries and default values.
+        """
+        default_engine, default_state = self.get_pdf_engine_state()
+        if default_state == 'install':
+            # pdf engine is not installed
+            # the call should be catched before (cf /report/check_pdf_engine) but
+            # if get_pdf is called manually (email template), the check could be
+            # bypassed
+            raise UserError(_("Unable to find '%s' on this system. The PDF can not be created.", default_engine))
+
+        engine, engine_state = self.get_pdf_engine_state(engine_name) if engine_name else (default_engine, default_state)
+
+        if engine_state == 'install':
+            # pdf engine is not installed
+            # the call should be catched before (cf /report/check_pdf_engine) but
+            # if get_pdf is called manually (email template), the check could be
+            # bypassed
+            _logger.warning(f"Unable to find {engine!r} on this system. The PDF will be created with {default_engine!r}.")
+            engine = default_engine
+
+        return getattr(self, f'_run_{engine}')
+
     def _render_qweb_pdf_prepare_streams(self, report_ref, data, res_ids=None):
         if not data:
             data = {}
@@ -823,19 +492,13 @@ class IrActionsReport(models.Model):
                     'attachment': attachment,
                 }
 
-        # Call 'wkhtmltopdf' to generate the missing streams.
+        # Call pdf engine to generate the missing streams.
         res_ids_wo_stream = [res_id for res_id, stream_data in collected_streams.items() if not stream_data['stream']]
         all_res_ids_wo_stream = res_ids if has_duplicated_ids else res_ids_wo_stream
-        is_whtmltopdf_needed = not res_ids or res_ids_wo_stream
+        is_pdfrendering_needed = not res_ids or res_ids_wo_stream
 
-        if is_whtmltopdf_needed:
-
-            if self.get_wkhtmltopdf_state() == 'install':
-                # wkhtmltopdf is not installed
-                # the call should be catched before (cf /report/check_wkhtmltopdf) but
-                # if get_pdf is called manually (email template), the check could be
-                # bypassed
-                raise UserError(_("Unable to find Wkhtmltopdf on this system. The PDF can not be created."))
+        if is_pdfrendering_needed:
+            pdf_engine = self._get_render_pdf_engine(data.get('report_subtype'))
 
             # Disable the debug mode in the PDF rendering in order to not split the assets bundle
             # into separated files to load. This is done because of an issue in wkhtmltopdf
@@ -857,16 +520,13 @@ class IrActionsReport(models.Model):
                     report_sudo.name,
                 ))
 
-            pdf_content = self._run_wkhtmltopdf(
-                bodies,
+            pdf_content_stream = io.BytesIO(pdf_engine(
                 report_ref=report_ref,
-                header=header,
-                footer=footer,
+                bodies=bodies, header=header, footer=footer,
                 landscape=self._context.get('landscape'),
-                specific_paperformat_args=specific_paperformat_args,
                 set_viewport_size=self._context.get('set_viewport_size'),
-            )
-            pdf_content_stream = io.BytesIO(pdf_content)
+                specific_paperformat_args=specific_paperformat_args,
+            ))
 
             # Printing a PDF report without any records. The content could be returned directly.
             if has_duplicated_ids or not res_ids:
