+++ odoo/addons/base/tests/test_split_table.py
@@ -1,6 +1,6 @@
 from odoo.tests.common import TransactionCase
 from odoo.tools import file_open
-from odoo.addons.base.models.ir_actions_report import _split_table
+from odoo.addons.base.models.ir_actions_report_wkhtmltopdf import _split_table
 from lxml import etree
 
 def cleanup_string(s):
